+++ gulpfile.js
@@ -74,6 +74,7 @@ gulp.task('doc', function(cb) {
         'src/ui/Dialog.js',
         'src/ui/Input.js',
         'src/ui/Page.js',
+        'src/ui/Panel.js',
         'src/ui/Radio.js',
         'src/ui/Tab.js',
         'src/ui/Textarea.js',